Added Loan Eligibility Prediction EDA Project
Loan Eligibility Prediction EDA Project: This notebook provides an in-depth exploratory data analysis (EDA) for a loan eligibility prediction project. It covers comprehensive data cleaning, visualization of key insights, and initial exploratory analysis to understand the dataset's characteristics. This EDA lays the foundation for subsequent stages of modeling and prediction in the loan eligibility domain.